# lib/api-zod/src/generated/types/createGoalBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateGoalBody {
  title: string;
  description?: string;
  category?: string;
  deadline?: string;
  successCriteria?: string;
}